backend/foodgram/requirements.txt
asgiref==3.6.0
atomicwrites==1.4.1
attrs==23.1.0
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
defusedxml==0.7.1
Django==3.2.3
django-debug-toolbar==3.2.4
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.13
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
exceptiongroup==1.1.3
Faker==12.0.1
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mixer==7.1.2
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
Pillow==10.0.0
platformdirs==3.5.0
pluggy==0.13.1
psycopg2-binary==2.9.7
py==1.11.0
pycparser==2.21
PyJWT==2.8.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
reportlab==4.0.4
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.1
sorl-thumbnail==12.7.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
